requirements.txt
bintrees==2.0.2
numpy>=1.12.1
pandas>=0.20.2
enum34>=1.1.6
tqdm>=4.32.1
gitpython>=2.1.8
TA-lib
matplotlib
future-fstrings
